requirements.txt
discord-py-interactions
openai